Fix can't multiply sequence by non-int of type 'numpy.float64' error

+++ openbb_terminal/stocks/options/hedge_model.py
@@ -15,7 +15,7 @@ def calc_hedge(portfolio_option_amount, side, greeks, sign):
 
     Parameters
     ----------
-    portfolio_option_amount: int
+    portfolio_option_amount: float
         Number to show
     side: str
         Whether you have a Call or Put instrument
